wip
Former-commit-id: 3f0dd8235a2970f4849778e627e8e4f81cda27a9

+++ src/my_size.h
@@ -96,7 +96,7 @@ public:
 };
 
 typedef my_asize<int> size;
-typedef my_asize<double> fsize;
+typedef my_asize<float> fsize;
 
 template<class T> class my_asize3d
 {
